requirements.txt
efficientnet-pytorch
kaggle
matplotlib
mediapipe
mlflow
numpy
opencv-python-headless
pandas
Pillow
PyYAML
python-multipart
pytest
scikit-learn
seaborn
streamlit
timm
torch>=2.0.0
torchvision>=0.15.0
tqdm
